Coming soon
Thay đổi hoàn toàn cơ chế di chuyển, bắn, delay

+++ src/towerdefense/gameEntity/gameTile/GameTile.java
@@ -10,7 +10,7 @@ public abstract class GameTile implements GameEntity{
 	public void setTexturePath(String s) {
 		this.texturePath=s;
 	}
-	public void setPos(int x, int y) {
+	public void setPos(float x, float y) {
 		pos.setPosition(x, y);
 	}
 }
